fix: address review findings — remove dead code, fix bugs, deduplicate
Bug fixes:
- Fix window.open in BillingTab (broken in Electron) → use openExternalUrl
- Fix division by zero when periodCredits is 0 (NaN CSS width)
- Fix background poll apiKey check not reactive to auth changes
- Add fallback for undefined creditsPerMin from API
- Fix variable shadowing in PaywallModal TIERS.map

Dead code removed:
- Remove unused createCheckoutSession, createPortalSession exports
- Remove unused openPortal, getInferenceToken from BillingContext
- Remove unused allRates state field and requestInferenceToken import
- Remove unwired TransitionBanner component
- Remove stale onOpenSettings/onOpenPlugins props from GraphToolbar,
  GraphEditor, and StreamPage

Deduplication:
- Extract RedeemCodeSection into shared component used by both
  PaywallModal and BillingTab
- Consolidate dashboard URL into single DASHBOARD_USAGE_URL constant
  in billing.ts
